[Mysql Client] Add QueryArgument for variant argument class
Previously, we used dynamic to hold the arguments. That proved to be a
bit slow and now that we want to add subqueries, it's impossible to use
it anymore.
This new class will accept Strings, ints, bool, double, null, vector of
itself, and Query's as arguments.

I maintained the Dynamic support since we have so many usages already,
but we can convert them once this diff gets in. All the constructors I
added are meant to be as close as possible to dynamic behaviour,
therefore making conversion easy. When a dynamic is passed, it is
converted to out storage.
Some of our queries require a pair or a list of pairs, they are
semantically the same. When QueryArgument is in Pair mode, it will
accept operator () to append values to it. Similar to dynamic::object,
but better because the order the user passes will be kept.