Improve DBEventLogger to get more data Squangle
Summary:
This remodels how to interact with DBEventLogger, now it became a template
and for Squangle it will receive a pair of ConnectionKey and context data.
But more information can be added.

With this extra data, this diff adds

 - tier_name for when in Async Mysql and passing DbLocator
 - db_hostname + mysql_hostname (When mysql is being proxied they may diverge and to keep consistent with the data we current have, using db_hostname)
 - db_ip

Any additional information can be added to
`ConnectionOperation` (used for Connect events) and then moved to
`Connection` (used for query events).

Some users don't set ConnectionParams fields for DatabaseHostname and EndpointHostname, thus we don't always have a hostname for the db, but we always have the IP. Since we don't want any DNS lookups, we are not doing any. If the data isn't given we will leave for a post-processing to deal with it.

Test Plan:
ptail -f perfpipe_dbclient_api_query

Running tao, squangle and olddb tests with asan

Bad formatted data, in case anyone wants to see
P19899582

Compiling HHVM

Reviewers: anca, chip

Reviewed By: chip

Subscribers: trunkagent, fbcode-common-diffs@, anca, lins, yfeldblum, chaoc

Differential Revision: https://phabricator.fb.com/D2194586

Signature: t1:2194586:1435707740:569ade1733f23e2393b145288db5f9a68a681e89