Support for enabling SSL in squangle by setting SSL_CTX
Summary:
Exposed setting an SSL_CTX to be used for the connection. Setting a
SSLContext is enough to enable the use of SSL for that connection.
Prefered to request a folly::SSLContext instead of a raw SSL_CTX
as it will insure the openssl thread library is correctly initialized (
done in folly::SSLContext constructor).

Test Plan:
Tested with pmysql2:
_bin/common/db/tools/pmysql2 -num_threads=100
-tier=dbproxy.cdb.read.oregon --port_offset=-10  "select 1"
--print_results=false

Reviewers: lins, chip

Reviewed By: lins, chip

Subscribers: fbcode-common-diffs@, ssl-diffs@, lins, yfeldblum, santoshb

Differential Revision: https://phabricator.fb.com/D2128498

Signature: t1:2128498:1433455485:d9de3a35680e17c1c2c199734ebf158ebab0264b