Improve visibility of operation timings in the client
Summary:
This adds last activity in the connection, the internal IO loop average
this is measured by EventBase) and average delay after scheduling a
callback.
I've changed lastSuccessfulQuery to lastActivity to simplify and have it
being the "connect time" too. I'm only updating it in Success cases
(successfully queried and connected) even though invalid SQL queries
also reset the idle time.
I'm exposing the stats data in the result so we have the context of that
operation. In order to expose it to PHP, I had to add to
AsyncMysqlConnection a ConnectResult that holds the operation pointer
(for timings)

This data will be used to better investigate the scenarios:
- You got a 2006? Here what to look for!
  - When was the last activity in the connection? This timestamp is put
    right when the query finishes, so IO thread is running it, if ASIO
    gets it only long after that timestamp, then we know Asio is guilty
    here.
  - If last activity is valid, check the delay average
- 2013? Let's check delay average and IO loop time avg.
  - Start time wasn't set for the connection based on the C++ side (in
    the extension there isn't a "result" for Connection, so PHP was
    timing it on its side). I added to the AsyncMysqlConnection the
    stats data so we can have accurate clear data in samples.
  - For this error the delay average is the most important piece of
    information. That and start time :)

PHP side has the idle timeout of the user, so I'll make some checks in
that side. It's hard to understand the scenarios by just trying to repro
it. I want this data in production so that I can use it to improve the
error messages/better diagnosis.

Extension part will be in a separate diff

Test Plan:
Testing HHVM with the new values and checking if they make sense

Reviewers: chip, anca

Reviewed By: anca

Subscribers: trunkagent, hphp-diffs@, fbcode-common-diffs@, jwatzman, anca, lins, jezng, yfeldblum, ebergen

Differential Revision: https://phabricator.fb.com/D2234578

Tasks: 7732151

Signature: t1:2234578:1437108241:cea66b8f8fe15d64b1a8800da2022cb4a0d3b48d