Implementing Async MySQL Row fetch streaming - Part 1
Summary:Allow callbacks during a fetch to pause the operation and allow another thread the consume the stream, therefore avoiding copy of the buffer.

To allow no copy, I added a new family of classes "Ephemeral" ones. For now, `EphemeralRowFields` and `EphemeralRow` are with less features  than buffered version to simplify and keep this diff with less noise. They directly encapsulate C MYSQL layer pointers and buffers.

All Ephemeral classes are only alive for the duration of the fetch, they shouldn't be copied, etc. (Later on, we can have the regular callback be this ephemeral class and if they copy the row or move, we buffer it)

`FetchOperation` used to keep a copy of the result as rows were being fetched.
In this diff `QueryOperation` and `MultiQueryOperation` now consume `RowStream` and creates the RowBlock themselves.

The `specialized` calls now are `notify` because that's more appropriate name for them.

Reviewed By: anca-agape

Differential Revision: D3143670

fb-gh-sync-id: 3449dc4dca38e2a70b871e17aed6df54d7da5447
fbshipit-source-id: 3449dc4dca38e2a70b871e17aed6df54d7da5447