Adds Pause in fetchOperation for Row Streaming - Part 2
Summary:
**Goal:**
Allow callbacks during a fetch to pause the operation and allow another thread the consume the stream, therefore avoiding copy of the buffer.

**What's happening here**
**Pause Implementation**
All pause logic is inside FetchOperation, this class is parent of QueryOperation and MultiQueryOperation.
`QueryOperation` and `MultiQueryOperation` handle buffered results, so this adds `MultiQueryStreamOperation` as subclass of `FetchOperation` that adapts the result directly to Stream callbacks. During stream callbacks, the user can pause the operation and transfer the power to another thread to read directly from underlying layers.

FetchOperation controls the state machine of a mysql query, as well as "loop the socket" until we have rows that the user can read.

`pause` can be called during a callback for StreamState: InitQuery, RowsReady, QueryEnded; When `pause` is called, we change the fetch_action to be `WaitForConsumer` and cancel the timeout.
Why is timeout cancelled? Because when we pause, we shouldn't have the IO Thread trying to access the Operation at all, we should respect the user's request, when `resume` is called, then timeout will be processed.

To track state `RowStream` was added, this is the abstraction that the consumer will be given to read rows and that will be used to inform the FetchOperation that there are no more rows to be read.

Reviewed By: anca-agape

Differential Revision: D3188747

fb-gh-sync-id: 1452bb7d5a752172c5b27663d592461829fc5543
fbshipit-source-id: 1452bb7d5a752172c5b27663d592461829fc5543