Add Fiber streaming API for MultiQueries
Summary:
**What**
This diff adds a new API using multi thread streaming to read the result without copies and with nice fiber interface.
MultiQueries are the more complex and complete use case, so that's the one we are starting with `stream_multiQuery` receives the query and returns the stream handler.

**Why**
Streaming properly is complex and requires a lot of synchronization between the user's thread and the IO thread. This API aims to allow a "sync" looking code, Fibers, to take advantage of no-copy reads.

**How**
Using Baton's and FetchOperation::pauseForConsumer/resume we are able to create a lock-free synchronization allowing user's thread to read the result and then unblocking the IO process.

**Synchronization flow:**
Before the user thread can do anything, it will get blocked on a `baton`, that baton will only be freed when the IO Thread is paused or ended.
E.g:
Initial scenario:
the user calls `streamMultiQuery` and calls `nextQueryResult`:
1- Before we can allow the user to have a `StreamedQueryResult`, we need to know if that result will start or fail. So the user will `wait` on the state baton;
2- Once the IO thread callback comes, depending on the callback reason (no-rows, rows, failure) a result and state will be set in `MultiQueryStreamHandler` and in the `StreamedQueryResult` then pausing the IO thread and freeing the baton.
3- The user thread is now free to read its state and receive the result.

For all the interactions the rule is the user is either waiting on a baton and IO thread is working, or baton is posted and IO thread work is paused.

**Entities:**

`MultiQueryStreamHandler` - Receives stream callbacks to control synchronization for empty results and wait between them. As well as protecting the access to the `Connection`. The user shouldn't `releaseConnection` until everything has ended.

`StreamedQueryResult` - Handles the fetching synchronization and provides iterator interface for rows.

Reviewed By: anca-agape

Differential Revision: D3183842

fb-gh-sync-id: f8311313e2535ff4b42aa13bfd66ac0e5ff4271c
fbshipit-source-id: f8311313e2535ff4b42aa13bfd66ac0e5ff4271c