Avoid mallocs for MultiQueryStreamHandler and StreamedQueryResult
Summary:
Avoid mallocs by Creating MultiQueryStreamHandler on stack and moving it out.
Consequently, we will need the user to invoke start on the handler to schedule
the operation to be run on the async thread and kick off the state machine.

For StreamedQueryResult we return an Optional to avoid mallocs

Reviewed By: drilibo

Differential Revision: D3510888

fbshipit-source-id: b7b716b82802abc6866049503e312e018b8aa150