Optimize query generation
Summary: Avoid copies completely for unsafe queries. The overall idea is that if the caller can guarantee the life time of the sql stmt, they Query object can just store the StringPiece passed. Else we make copy and create a StringPiece on that copy. However this is managed transparently by the QueryText and outside of that class one only calls getQuery() to get the StringPiece representing the query. This would allow us to avoid any copies when the query is an unsafe query and caller requested a shallow copy. The diff also adds MultiQuery class that wraps multiple queries and a rendered_query. Both QueryOperation and MultiQueryOperations use MultiQuery to store the queries.

Reviewed By: drilibo

Differential Revision: D3495772

fbshipit-source-id: 8dc327f42b04d860cd1be8db0c0b8463584850b7