Introduce SyncMysqlClient and SyncConnection for inline processing
Summary:
- Why
In order to have efficient usage of AsyncMysqlClient in 1 thread, we need to run the loop inline with the rest of execution. The simplest approach is to create separate classes for synchronous handling, this eliminates a lot of overhead from AsyncMysqlClient and gives us more flexibility in memory allocation and thread handling.

- How

Added a separate Client to wrap the logic we need for Sync and remove the overhead for Async. SyncMysqlClient always calls `loop` when `runInThread` is called. This is being done because some classes as (MysqlConnectionHolder) call runInThread.

`Connection` is already used as EventHandler and synchronization for operations, so I basically added subclasses to differentiate between baton’s and `loop`.

Remove dependency on baton’s, `MultiQueryStreamHandler` minimum changes so far to just use synchronization methods defined in `Connection`. So in Fiber mode, we will use AsyncConnection, which is baton.

Reviewed By: sarangbh

Differential Revision: D3523998

fbshipit-source-id: 83e34bbfe902a3aea2e728f895d4e803c0ade7bd