Change fbstring& constructor etc to use StringPiece
Summary:
Should be transparent to callers which were using `fbstring` before. Means that callers who put the names they want to use into `constexpr StringPiece` constants don't need to call `.str()` all over the place.

It's not possible to have both because the compiler can't resolve which one to use.

Reviewed By: yfeldblum

Differential Revision: D3605203

fbshipit-source-id: 728c57dce5f15f98c88c6f7dcb9493507e0a4ce1