Fix compilation under GCC 4.8
Summary:
The open source build of HHVM still needs to be able to be compiled with GCC 4.8, and GCC 4.8 was complaining about syntax errors, as it doesn't seem to like attempts to expand a variadic parameter pack as a capture in a lambda.
This just switches it to use `std::bind` instead.

Reviewed By: drilibo

Differential Revision: D3699453

fbshipit-source-id: 9f5331c6bbcba0cf3089369eb8d529c82aed65ff