Allow qualifying and aliasing column names in SELECT statements.
Summary: When constructing a query column names are blindly wrapped in back ticks. This patch allows a two- or three-element dynamic::array to be passed in as a QueryArgument for both '%C' and '%LC' substitutions to either qualify column names with table names or both qualify column names and alias them (using an 'AS' clause).

Reviewed By: drilibo

Differential Revision: D3692137

fbshipit-source-id: c5eca3d3001dac868fa1f4a1fd77e9d5e0d15182