Improve timeout cause detection
Summary:
In certain cases, we may timeout because of an overloaded/stalled
EventBase (Maybe someone blocked inside a future callback, or a large block of
data is being decompressed/decrypted, etc). We want to detect
as many situations as possible where timeouts are caused by thread
overload, rather than a true network timeout and pass that information along to
the user and logging if we can.

(D3888404 seems to be a duplicate of this diff for some reason)

Reviewed By: abal147

Differential Revision: D3889067

fbshipit-source-id: b7797b414765dcf38142a551d3a2da02b6fab30e