much lower default timeout
Summary: 60 seconds is ridiculous for a default, as very few things even have an application timeout that high. Set connect to 1.03 seconds (0.03 to be able to handle a tcp syn rto in region), and total to 3 seconds, which is more in line with typical defaults. See https://our.intern.facebook.com/intern/sevmanager/view/s/143980/ for more context on how super long timeouts like this can cause things to go south

Reviewed By: abal147

Differential Revision: D4464677

fbshipit-source-id: 605552d224d4a31b64d67764220ff976cf71621b