Set sendKillOnQueryTimeout() to true by default when not using a
Summary:
For connections through a proxy, the proxy handles query timeouts and
kills queries that have exceeded their timeouts. For direct connections we do
not current send the kill on timeout, unless the option is explicitly set. Here
we set the value to true if not using a proxy, and false if so automatically.

Reviewed By: anca-agape

Differential Revision: D4656280

fbshipit-source-id: 2cec473a196d19c11a11e6358c50053acd597dd7