Add Query size in bytes to result sets
Summary:
Adds the size in bytes. This is limited to the application
row bytes as that's all we can see. There are ways to get this value
closer to the actual bytes sent over network but this is a close enough
proxy. Unless we need to deal with bad actors who name their columns
very long names this is sufficient for identifying queries with large
result sets

Reviewed By: jrahman

Differential Revision: D5513015

fbshipit-source-id: 1d32bad02864c230c8857883d7995c517ff0e810