Don't run freed Operations
Summary:
Looks like the async pool has a known race where
the operation might be cancelled before it first runs on
the async client. I suspect this was mostly expected during
shutdown cases.

In T22685228 we are seeing seg faults in this code because
the AsyncClient cleanupCompletedOperations() is running before
the first run of the operation. Since a raw pointer is captured
in the lambda, it seg faults on first access into the object.

Fix is to grab a weak ptr and check if the object is alive before running
other operations.

Differential Revision: D6214506

fbshipit-source-id: 1831aaa93161b299b4f8d5b94b6ac0ce3df06a84