Add comment about the use of mysql_free_result
Summary:
See comment.
Originally it was thought that we should eventually move to
mysql_free_result_nonblocking. This is infeasible due to the way
that we call mysql_free_result currently, but this works correctly
for our use case.

Ultimately mysql_free_result and the nonblocking versions will do
the exact same thing. The justification for using mysql_free_result
is that it would look very odd to call a reentrant function as a
destructor, even though it also looks weird to use sync function
here

Reviewed By: lth

Differential Revision: D6539033

fbshipit-source-id: f4e23448a755260446306152a151ed8d60d29497