Change mustSucceed() to throw rather than LOG(FATAL).
Summary:
The mustSucceed() method for ConnectOperation and QueryOperation is used in testing to make sure that the requested operation succeeds.  If it does not, the function called LOG(FATAL) which generated a SIGABRT (which has been okay in the past since this function is intended for tests only).  However, this is not very flexible.  It would be preferable to have it throw an exception giving callers the option of catching it, allowing them to choose their desired behavior.

I changed the code to throw a new exception - facebook::db::RequiredOperationFailedException.  Existing applications will still stop, just with an uncaught exception instead of a fatal log message.

Reviewed By: jrahman

Differential Revision: D6669248

fbshipit-source-id: 63f2c9f345a4410b0970efb0c1495a647237e7ed