Get better errors from ConnectAttempts
Summary:
I've seen two different errors in AsyncMysqlTest.ConnectAttempts recently, but the errors aren't sufficient to tell what went wrong.  Update the code to give better information for the future.

Also added an operator<<(..., OperationResult) so that when the test fails we will get the enumeration name, not just the byte pattern.

Reviewed By: jrahman

Differential Revision: D7417059

fbshipit-source-id: 2f23e57e2406a1fba6dbcfb7fed358bed54aaa21