Change pointer to reference (3/7)
Summary:
The dbclient APIs that take a DbLocator originally took a pointer.  This was confusing and some clients felt they needed to allocate DbLocator objects separately.  The API has been updated to take a const reference and the old pointer API will be removed soon.

This diff updates a portion of the callsites to use the new API that takes references.

Reviewed By: jrahman

Differential Revision: D18773318

fbshipit-source-id: fc23cf5270b41b93025d0a0592fc6a4460e189b1