Revamp ODS for MySQL clients
Summary:
Add new ODS logging for mysql clients. The changes are going to be introduced in phases, and this diff only has the C++ implementation (python/php/java and reporting/alarms/dashboards to follow).

The design doc is here: https://fb.quip.com/wB4xAvawxObq

The new ODS entries will have the following identities:

  entity: "database_access.<LANGUAGE>.<REGION>.<SHARDMAP>.<ENDPOINT_TYPE>.<RESULT>.<ERRNO>",
  key: "<opened_connections | closed_connections | queries>.sum.60",

Currently the new ODS logging is controlled by feature gating (currently fully disabled, and will slowly be enabled over time). While the rollout is happening, the old ODS logs are still being created, to not break any reporting/dashboards. When fully rolled out, the old ODS logs will be removed from the codebase.

NOTE: The design doc does not include <LANGUAGE> and <ERRNO> as dimensions, but after some discussions, the number of timeseries that will be created by adding these 2 dimensions won't be an issue (Should be between, 10-20k, which was determined to be acceptable)

Reviewed By: jkedgar

Differential Revision: D22883738

fbshipit-source-id: 3fdefd32a62087ce91e8c1e79c8bbcaa5e10e4dd