Add tcp handshake connect timeout
Summary:
Adding tcp handshake timeout for connect, so that clients have a way to fail fast in case the latency in a region is high instead of waiting for the entire connect timeout.

Tcp timeout covers the time take to complete upto CONNECT_STAGE_NET_COMPLETE_CONNECT. That is the phase by which the initial tcp socket connection is established and before we get the intial response from the server. (https://dev.mysql.com/doc/internals/en/plain-handshake.html).

Once the server moves to CONNECT_STAGE_READ_GREETING stage, the timeout is cancelled, and is a no-op.

Reviewed By: jkedgar

Differential Revision: D23585508

fbshipit-source-id: bace2c03361e160708dbcdcbc10e4c9dec41e8af