Refactor olddb
Summary:
Some refactoring in olddb to make the changes in TAO easier.  The connect function now has two different levels of calls - a high level connectImpl() and a lower level blockingConnect().  This is similar to how the Query function is implemented.  This is needed because the low-level functions take a MYSQL* and I need to remove the requirement of having a MYSQL*.

This also makes escapeString() a virtual function (to be overridden in a subsequent diff) as well as provide methods to get error information from DBResults.

Reviewed By: yichenshen

Differential Revision: D23980338

fbshipit-source-id: a1bd5b29b2734a53f050b5d9256a8c2ac5c92765