Support StringPiece parameter for escapeString()
Summary: As part of the work for migrating TAO off of olddb, I discovered that TAO needs to call escapeString with an existing `const char*` and `size_t` (which is supported in olddb).  We should not need to build a new string just to call `escapeString`, so this modifies `escapeString` to also support a folly::StringPiece as the parameter.

Reviewed By: yfeldblum

Differential Revision: D24691165

fbshipit-source-id: b036af16d58c483755a1cbb5be1fbcb931e9f934