Migrate MySQL to AsyncSSLSocket::getSSLSessionV2
Summary:
SSLOptionsProviderBase is used for two types of connections. Client -> myrouter connections, which use SSL_SESSION for resumption, and myrouter -> backend connections, which use AsyncSSLSocket and therefore folly::ssl::SSLSession with the new API.

Both can be supported by caching a variant of the two types.

Alternatives Considered:
1. Storing folly::ssl::SSLSession in ThreadSafeSSLSessionCache, creating OpenSSLSession objects directly from raw SSL_SESSION.
This exposes the OpenSSLSession class, which we'd rather keep an implementation detail, and forces unnecessary conversions for SSL_SESSION <-> OpenSSLSession for mysql client resumption.

Reviewed By: mingtaoy

Differential Revision: D22038802

fbshipit-source-id: c23006597ff98714a7444033f4e22a55fa34573a