Send COM_RESET_CONNECTION before recycling idle connection
Summary:
This diff will make sure that each idle connection sends COM_RESET_CONNECTION before returning back to client-side connection pool. This will (1) make the connection clean by resetting all dirty session variables, to make sure that connections in the pool can be reused with a clean state, and (2) make MyRouter, in response to COM_RESET_CONNECTION command, detach backend connections so that they can be resued in the MyRouter backend connection pool.

recycleMysqlConnection(), a callback function in charge of recycling idle connection, is called by destructor of Connection object, so by the time this function is called, the corresponding Connection object is being destroyed. Thus, in order to send another command to MySQL instance with the connection, we need to clone Connection objest that is being destroyed. Also, I added a new Operation class, ResetOperation, which is responsible for sending COM_RESET_CONNECTION commmand and waiting for OK packet from MySQL instance (or MyRouter instance).

(TODO with the next diff)
Add a gating logic to roll this feature out gradually, with a new connection parameter)

Reviewed By: aditya-jalan

Differential Revision: D25688704

fbshipit-source-id: 311c9c4c027ee27b64550ac03ea4980ea53f38df