Remove implicit `dynamic` constructor from squangle QueryArgument
Summary:
This was causing issues for an upcoming change that allows
implicitly making dynamic from anything 'string-like' including
StringPiece, as this then caused ambiguous overloads to happen when
constructing QueryArgument. Seems like this functionality is 'legacy'
but we can still handle this appropriately. And potentially make things
more efficient down the line once we identify who is relying on this...

First step, move the dynamic 'constructor' into a static factory function
fromDynamic, which is nicely bgs-able and removes it from the overload set.
However, we also have to handle the varargs constructor for Query which
forwards all arguments to vector<QueryArgument>::emplace_back, so add
static 'from' functions that delegate either to fromDynamic or the
normal constructor as appropriate, and call those from unpack.

I noticed while converting callsites that several places copy data into
the dynamic only for it to get copied back out again later. This is
grossly inefficient. Following up with diffs to support Range based
constructors to avoid these copies would be a good idea.

Reviewed By: yfeldblum

Differential Revision: D26257420

fbshipit-source-id: c0f6264ed511cb0d0be47d5f7bb72c6f76149818