Return error if configManager is unavailable
Summary:
Fail early when config manager is null for c++ client. As we are taking dependency on client config for determining imp conn/query properties, we do not want to connect without the right infomation. This makes sure we are not missing any important config.

ConfigManager is a singleton and can be null, if the code executes before static initializations, or some wierd scenario where singleton might have been released. These are all caused by wrong usage of lib, and should not affect any of actual usecases.

Reviewed By: jkedgar

Differential Revision: D25264532

fbshipit-source-id: 1356e1481f887929bc3743b6e1f23b941a962423