Mark read-only functions const noexcept
Summary: In Rust, the compiler enforces that if someone takes a mutable reference of an object, the reference must be exclusive. On the other hand, if we want multiple references to the same object, the reference must be immutable. This applies to C++ when we build Rust bindings that use C++ objects. It seems reasonable to be able to create shared references to `PoolStats` and they are free to call these `numXXX` functions at the same time. These functions do not modify any internal state and do not throw exceptions, so add `const noexcept`.

Reviewed By: dtolnay

Differential Revision: D26443555

fbshipit-source-id: 2c365f9b53bf6fd43bc4d3bedf865fcd28b755cc