reset connection before reusing idle connection, even when connection is destroyed from eventbase thread
Summary:
D25688704 (https://github.com/facebook/squangle/commit/f85d4f3b56a70843762621e3ca63303a2f250cd3) ensures that an idle connection is properly cleaned with COM_RESET_CONNECTION before it is returned back to connection pool, but it still had some constraints, one of which is that connection is reset only when it is destroyed from non-eventbase thread. This constraint was introduced because previously, reset connection happened in the destructor of connection object and destroying the object needs to be blocked until reset connection completes (because otherwise, destroyed connection cannot be reused immediately), but if the destructor happens to be called by eventbase thread, then we cannot block destroying the connection object (because it becomes deadlock).

In order to ensure that connection is cleaned even when it is destroyed from eventbase thread, this diff moves the moment when a connection is cleaned, from when a connection is destroyed to when a connection is fetched from a connection pool. This way, we don't need to block destructor and we can simply assign a callback function to the reset operation, and then it can complete the operation of making the connection ready for use. This is even better than the previous way because we reset connection only when a connection is about to be reused from a pool.

To do this, the destructor of connection marks a connection as `NeedResetBeforeReuse` if it is called by eventbase thread. Later, if a connection with this bit is fetched from a pool, then it is cleaned with COM_RESET_CONNECTION.

For now, I will let reset connection work this way only when a connection is destroyed from a eventbase thread. When this proves to be robust enough, then I will move ahead and let reset connection work this way all the time.

Reviewed By: jkedgar

Differential Revision: D26810439

fbshipit-source-id: ea9143d48b105df6cb84cf7bb843e00c7137389f