Don't move callbacks back and forth
Summary:
In the old method, callbacks are moved from the connection to operations and then back to the connection.  This makes it difficult to add callbacks for queries as they will get persisted in the connection.

If, instead, we only move callbacks once - when the connection operation finishes into the connection object - and have queries have an initial callback that calls the connections callbacks, we have the option to add more callbacks on the query that won't be persisted.

Reviewed By: jupyung

Differential Revision: D27281693

fbshipit-source-id: 7f68667a5e79e229557ebbcda507e5b5f936563f