Fix logging for old connection path
Summary:
The change over from Olddb to SyncMysqlClient logging (D25504654) was incomplete.  The FbSyncMysqlClient used in the new connection path was correctly updated to have a logger, but the SyncMysqlClient used in the old connection path was not.  This meant that when the old connection path was being used no log entries were ever sent to Scuba.

In addition, the code to turn off Olddb logging was missing a check.  All query logging should go through the SyncMysqlClient version but for the old connection path we should still do the olddb logging while the new connection path should use [Fb]SyncMysqlClient logging.

Added a new test to validate that there was an existing logger and to also verify that the logging calls get called correctly.  This involved moving some test functions up to the MysqlClientBase class.

Reviewed By: davidgoode

Differential Revision: D28206423

fbshipit-source-id: 6819cff08fd29659ce9a488c57d29a544a5ec7dc