Ensure that when connection operation hits timeout, it cancels the related reset operation
Summary: ConnectPoolOperation may have its dependent pre-operation, an operation that is required before completing it, which could be reset_connection or change_user (not implemented yet) operation. ConnectPoolOperation is completed only after its pre-operation is completed, if there is. This diff ensures that when ConnectPoolOperation hits timeout, it cancels the pre-operation.

Reviewed By: jkedgar

Differential Revision: D28346840

fbshipit-source-id: c344afa5d926c01ff68e1cfbf0afbca07f6bfecb