Handle protect escapeString
Summary:
We've looked at multiple ways of protecting TAO's use of escapeString and this is the current simplest one.  The issue is that TAO uses connections in multiple threads simultaneously in a single case - the use of escapeString.  This worked in olddb because the code just used a raw connection pointer which was always present.  In the new code, the client moves the connection as a unique_ptr in and out of query operations and if TAO calls escapeString right as the connection is moving we can get bad behavior (dereferencing null pointers).

We looked at modifying TAO not to have multi-threaded access to escapeString while a connection is being used elsewhere, but it is not trivial to achieve.  The long term goal _is_ to have TAO change, but in the meantime we are adding a mutex in MySqlConnectionOnAsyncApi.  The mutex is _only_ held while getting the connection pointer prior to calling escapeString and when moving the connection into or out of the handler.  This means the mutex will almost always be un-contended and thus very efficient but will still protect against the rare possibility that escapeString is called right as the connection is moving.

The one drawback to this method is that I needed to have a way to expose the connection pointer from the MultiQueryStreamHandler class.  This unfortunately would make it possible for someone else to abuse the code.  Fortunately, this class is rarely used.  According to this (https://fburl.com/codesearch/5ecl93zj), TAO appears to be the only user of it right now.

In addition with the investigation into the escapeString problem that we did we realized that escapeString should never return an empty string on an error.  This always worked in the past because escapeString _never_ failed, but TAO does not check the return value so accidentally returning an empty string could cause data loss as this is used on database changes and returning an empty string when escapeString was called column contents means that the data would be just lost.  Thus I changed it to throw an exception on failure.

Reviewed By: fadimounir

Differential Revision: D28425834

fbshipit-source-id: 9b215d37cf5a5d7162511901038899c04397149f