Add FOLLY_NODISCARD for methods returning SemiFutures
Summary:
Getting a `folly::SemiFuture` back from a function and not assigning it to an executor or calling `.get()` on it is a logic error that produces undesired results.

To save future heartache, let's mark all relevant returns as `FOLLY_NODISCARD` so that they will give compile errors instead of silently not working (or worse, working wrong).

Reviewed By: jkedgar

Differential Revision: D29036808

fbshipit-source-id: 798c185dc7a4314ee429af3a37f7b33f2b4a0f12