New classes that don't depend on olddb
Summary:
Now that everything is ready, introduce new classes to handle connections and datasets that don't rely on olddb whatsoever.  The code in the TaoMysql* classes can switch between the old and new ways of doing things.  The connection cache cleanup is in the next diff.

Much of the code is pulled from olddb and modernized.

Reviewed By: davidgoode

Differential Revision: D25505927

fbshipit-source-id: 547cec8e0d26685c03d930ac5c03e1bee663bcc7