Fix a rare case where pre-operation (change_user or reset_conn) is not cancelled when connection operation hits timeout
Summary:
We previously introduced pre-operation to ConnectionPoolOperation, to ensure that when the operation hits timeout, its dependent operation, like change_user or reset_conn, named "pre-operation" is also canceled. But even after the fix, a rare occurrence was similar events has been reported.

Looking at the coredump log, it turned out that the callback function of pre-operation called `ConnectPoolOperation::connectionCallback` while it already hits timeout; its mysql_errno_= 7004 (timed out error in pool). And its preOperation_ was not nullptr, meaning that its pre-operation was not canceled for some reason.

Digging deeper, there seems to be a rare case where `reuseConnWithChangeUser()` or `resetConnection()` is run very slow, so `rawPoolOp->setPreOperation()` is called later than the timeout of rawPoolOp (ConnectionPoolOperation). In that case, pre-operation is set AFTER it is canceled by timeout, so still pre-operation remains not cancelled.

In order to fix this, I introduced one more variable to PreOperation, canceled_. If pre-operation is canceled by timeout, it is set to true, and if other pre-operation tries to set this with `PreOperation::set()`, it fails and returning false. And then, `reuseConnWithChangeUser()` or `resetConnection()` does not try to run a pre-operation and quits (because its dependent connection pool operation already times out).

Reviewed By: jkedgar

Differential Revision: D31000493

fbshipit-source-id: f9e94e715cbd77bb30d3aa89a0180f68d71ec81d