Provide a bit of type hint information on obscure error message
Summary:
When working with the Query class, you can occasionally hit the obtuse error message: "Only allowed type conversions are Int, Double, Bool and String" which can be somewhat confusing.  Add code to include the type that the compiler is trying to convert to potentially help future coders.

I got this output:

```
 Only allowed type conversions are Int, Double, Bool and String: type found: std::vector<facebook::common::mysql_client::QueryArgument, std::allocator<facebook::common::mysql_client::QueryArgument> >
```

Reviewed By: abal147

Differential Revision: D32503719

fbshipit-source-id: 01399cab6ccf197a96a6770a56903b5723749410