Fix a rare segfault happening when a pool is destroyed in the middle of running connection cleanup
Summary:
This diff fixes a rare segfault, which could occurs when an async conn pool is destroyed in the middle of running cleanupConnections().
Because cleanupConnections() may delete MySQLPooledHolder from its map, decreasing reference count of shared_ptr<AsyncConnectionPool>, the shared pointer's reference count can become 0 if the pool is already deleted in the user context, then the pool's destructor could be called in the middle of cleanupConnections(). Because the pool's destructor cleans up maps that cleanupConnections() is traversing in, this can certainly lead to segfault.

This diff prevents this by cleanupConnections() preserving one unique_ptr<MysqlPooledHolder> pointer to prevent its pool from being destructed in the middle of traversing map. Also I changed the order of operations in timeoutExpired(), making cleanupConnections() run the last, so that if pool is ever destructed, it happens when timeoutExpired() is finished.

Reviewed By: jkedgar

Differential Revision: D32812146

fbshipit-source-id: cc97a726efd299eb34095952384d9448aa50a2b6