simplify the access to semaphore protected variables
Summary:
The pending_operations_mutex_ mutex was used to protect three different member variables.  This kind of code can be confusing and error-prone because nothing forces the mutex to be acquired before accessing the variables.

This diff switches the code to use folly::Synchronized and pulls the three into a structure so that they physically cannot be accessed without a lock.

Reviewed By: abal147

Differential Revision: D33487204

fbshipit-source-id: 44b9a2dfcfac4eba6456608e92973fc74fb7070c